bindsym --release
The functionality of this in sway is slightly different than i3. In sway
you can hold down the keys indefinately, then release to trigger the
command. In i3 if you hold down the keys the command will trigger after
a short amount of time. I don't use this feature so I am not sure if it
makes a difference to anyone